chore: move registerDecorators into component
