+++ packages/@lwc/babel-plugin-component/src/post-process/index.js
@@ -4,14 +4,12 @@
  * SPDX-License-Identifier: MIT
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
-const transform = require('./transform');
 const dedupeImports = require('./dedupe-imports');
 
 function exit(api) {
     return {
         Program: {
-            exit(path, state) {
-                path.traverse(transform(api), state);
+            exit(path) {
                 dedupeImports(api, path);
             },
         },
